backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.27.0
alembic==1.12.1
jinja2==3.1.2
